RAG/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
langchain==0.3.0
langchain-community==0.3.0
langchain-groq==0.2.0
faiss-cpu==1.13.2
pypdf==4.3.1
python-dotenv==1.0.1
sentence-transformers==3.0.1
